+++ apps/trackers/apps.py
@@ -1,0 +1,9 @@
+"""
+Trackers - everything related to the vulnerability trackers.
+"""
+
+from django.apps import AppConfig
+
+
+class Trackers(AppConfig):
+    name = "apps.trackers"